r1562: Cast into sbyte to get -1 selected index https://www.ascensiongamedev.com/bug_tracker/intersect/map-names-causes-crash-in-visual-interface-r1562/
Former-commit-id: 4c44140594e88f80e235c3ba76025ead1e565851

+++ Intersect.Editor/Forms/Editors/Events/CommandPrinter.cs
@@ -630,7 +630,7 @@ namespace Intersect.Editor.Forms.Editors.Events
                     if (MapList.GetOrderedMaps()[i].MapId == command.MapId)
                     {
                         return Strings.EventCommandList.playanimation.ToString(AnimationBase.GetName(command.AnimationId),Strings.EventCommandList.animationonmap.ToString(
-                                MapList.GetOrderedMaps()[i].Name, command.X, command.Y,Strings.Directions.dir[command.Dir]));
+                                MapList.GetOrderedMaps()[i].Name, command.X, command.Y,Strings.Directions.dir[(sbyte)command.Dir]));
                     }
                 }
                 return Strings.EventCommandList.playanimation.ToString(AnimationBase.GetName(command.AnimationId), Strings.EventCommandList.animationonmap.ToString(
